# Cuda optimizations and other loose ends