--- a/cuda/kernel_adam.cu
+++ b/cuda/kernel_adam.cu
@@ -77,7 +77,8 @@ gpu_gradient_minAdam_kernel(
 	__shared__ int entity_id;
 	__shared__ float best_energy;
 	__shared__ float sFloatAccumulator;
-	extern __shared__ float sFloatBuff[];
+
+	volatile __shared__ float sFloatBuff[6 * MAX_NUM_OF_ATOMS + 5 * ACTUAL_GENOTYPE_LENGTH];
 
 	// Ligand-atom position and partial energies
 	float3* calc_coords = (float3*)sFloatBuff;
@@ -130,7 +131,6 @@ gpu_gradient_minAdam_kernel(
 		#endif
 	}
 	__syncthreads();
-	__threadfence();
 	energy = pMem_energies_next[run_id * cData.dockpars.pop_size + entity_id];    
 
 	int offset = (run_id * cData.dockpars.pop_size + entity_id) * GENOTYPE_LENGTH_IN_GLOBMEM;
@@ -158,7 +158,6 @@ gpu_gradient_minAdam_kernel(
 	// E.g. in steepest descent "delta" is -1.0 * stepsize * gradient
 
 	// Asynchronous copy should be finished by here
-	__threadfence();
 	__syncthreads();
 
 	// Enable this for debugging ADADELTA from a defined initial genotype
@@ -210,7 +209,6 @@ gpu_gradient_minAdam_kernel(
 		// =============================================================
 		// =============================================================
 		// Calculating energy & gradient
-		__threadfence();
 		__syncthreads();
 
 		gpu_calc_energrad(
@@ -262,7 +260,6 @@ gpu_gradient_minAdam_kernel(
 			printf("\n");
 			#endif
 		}
-		__threadfence();
 		__syncthreads();
 		#endif // DEBUG_ENERGY_ADADELTA
 
@@ -299,7 +296,6 @@ gpu_gradient_minAdam_kernel(
 			// Applying update
 			genotype[i] -= mp / (sqrt(vp) + cData.dockpars.adam_epsilon);
 		}
-		__threadfence();
 		__syncthreads();
 
 		#if defined (DEBUG_SQDELTA_ADADELTA)
@@ -312,7 +308,6 @@ gpu_gradient_minAdam_kernel(
 				printf("%13u %20.6f %15.6f %15.6f %15.6f\n", i, square_gradient[i], delta[i], square_delta[i], genotype[i]);
 			}
 		}
-		__threadfence();
 		__syncthreads();
 		#endif
 
@@ -358,8 +353,7 @@ gpu_gradient_minAdam_kernel(
 			}
 #endif
 		}
-		__threadfence();
-		__syncthreads(); // making sure that iteration_cnt is up-to-date
+			__syncthreads(); // making sure that iteration_cnt is up-to-date
 #ifdef AD_RHO_CRITERION
 	} while ((iteration_cnt < cData.dockpars.max_num_of_iters)  && (rho > 0.01f));
 #else
@@ -378,7 +372,6 @@ gpu_gradient_minAdam_kernel(
 	}
 
 	// Updating old offspring in population
-	__threadfence();
 	__syncthreads();
 
 	offset = (run_id * cData.dockpars.pop_size + entity_id) * GENOTYPE_LENGTH_IN_GLOBMEM;
@@ -410,8 +403,7 @@ void gpu_gradient_minAdam(
                           float* pMem_energies_next
 )
 {
-	size_t sz_shared = (6 * MAX_NUM_OF_ATOMS + 5 * ACTUAL_GENOTYPE_LENGTH) * sizeof(float);
-	gpu_gradient_minAdam_kernel<<<blocks, threads, sz_shared>>>(pMem_conformations_next, pMem_energies_next);
+	gpu_gradient_minAdam_kernel<<<blocks, threads>>>(pMem_conformations_next, pMem_energies_next);
 	LAUNCHERROR("gpu_gradient_minAdam_kernel");
 #if 0
 	cudaError_t status;