… cuda/auxiliary_genetic.cu