This PR improves the performance of the Cuda path to finally be on par with the OpenCL version.

This was achieved by eliminating unnecessary synchronizations across blocks and by using static shared memory (enabled by the v1.5 optimizations) instead of dynamic shared memory.

Docking times comparison with local search method specified and otherwise default settings for rand-0 ligand of Diogo's 42 dataset on an Nvidia Titan V:

**Solis-Wets**
| NUMWI | OpenCL v1.5 | Cuda v1.5 | this PR |
|--------:|:-------------:|:---------:|:-------:|
|         64 |              143 s |       209 s |    143 s |
|        128 |              128 s |       226 s |   139 s |

**Adadelta**
| NUMWI | OpenCL v1.5 | Cuda v1.5 | this PR |
|--------:|:-------------:|:---------:|:-------:|
|         64 |              263 s |      295 s |    257 s |
|        128 |              221 s |      282 s |    233 s |

Please do a full set of test runs for both Solis-Wets and Adadelta.