@diogomart 2021-08-25
E50s are looking good for CUDA, ADADELTA, 128wi, using filelist and overlap=on, single-GPU.
![09d6579-ocl-128wi-ad_vs_PR151-107fd7a-cuda-128wi-overlap](https://user-images.githubusercontent.com/1564797/130740726-6c21d251-3c28-42a0-980b-9d80818f91a7.png)