@diogomart 2021-09-01
Running on 8 GPUs simultaneously also works :) This is good to merge :+1: 
![adgpu-1 5 0-cuda-128wi-overlap-sw_vs_PR151-f8fd68f-cuda-128wi-overlap-allgpu](https://user-images.githubusercontent.com/1564797/131715490-d7e3f7ce-8b1d-4d40-a2a6-b19e7cfde80a.png)